Updated upload command to use commit range
You can now specify the commit range to upload as an argument to
the 'upload' command. The range format is in the same parsing style
as libgit. eg.

$ nirbt upload HEAD..HEAD~3

will concat commit messages from HEAD~3 to HEAD to use as the
description. The diff will be taken from the latest commit through
to the parent of the oldest commit. eg. HEAD~3^, from the above
example.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>